raspberry_code/requirements.txt
awsiotsdk==1.5.14
awscrt==0.11.14
picamera==1.13
tensorflow==2.4.0
awsiot==0.1.3
pyserial==3.5
